fix approve campaign, getRakingRatio

+++ SWallet.Repository/Payload/Charts/TitleAdminModel.cs
@@ -9,7 +9,7 @@ namespace SWallet.Repository.Payload.Charts
     public class TitleAdminModel
     {
         public long? NumberOfBrands { get; set; }
-        public long? NumberOfCampagins { get; set; }
+        public int? NumberOfCampagins { get; set; }
         public long? NumberOfStudents { get; set; }
         public long? NumberOfProducts { get; set; }
     }
